requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
discord.py==2.6.4
frozenlist==1.8.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.11
multidict==6.7.0
oauthlib==3.3.1
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
urllib3==2.5.0
yarl==1.22.0
